templates/poetry2nix/pyproject.toml
[tool.poetry]
name = "imgapp"
version = "0.1.0"
description = ""
authors = ["Igor Barale <svitaxiom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
flask = "^3.0.0"
pillow = "^10.1.0"
numpy = "^1.26.2"
pandas = "^2.1.3"
torch = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
imgapp = "imgapp:main"
